flight-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
